Add configuration for recording time
Allow a user to configure the recording time for top, allowing for aggregation of requests longer than 5 seconds.

Also added a check in the ServiceProvider to only register event listeners when a valid recording time is set. This provides the added bonus disable this package by setting the recording time to 0.